Extract Interep routes in own controller
