+++ src/services/http/controllers/index.ts
@@ -1,4 +1,5 @@
 export * from './Events';
+export * from './Interep';
 export * from './Posts';
 export * from './Tags';
 export * from './Users';
